+++ tests/soho-tag/soho-tag.spec.ts
@@ -1,0 +1,25 @@
+import AxeBuilder from '@axe-core/playwright';
+import percySnapshot from '@percy/playwright';
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+test.describe('Tag tests', () => {
+  const url = '/ids-enterprise-ng-demo/tags';
+
+  test.beforeEach(async ({ page }) => {
+    await page.goto(url);
+  });
+
+  test.describe('general page checks', () => {
+    test('should have a title', async ({ page }) => {
+      await expect(page).toHaveTitle('IDS Enterprise - Angular Components');
+    });
+  });
+
+  test.describe('snapshot tests', () => {
+    test('should match the visual snapshot in percy', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      await percySnapshot(page, 'soho-tag-light');
+    });
+  });
+});